CATTY-79 Add support for Xcode 11 and Swift 5.1

+++ src/Catty/DataModel/Formula/FormulaElement.m
@@ -133,7 +133,7 @@
 - (NSString*)stringForElementType:(ElementType)type
 {
     NSDictionary *dict = kstringElementTypeDict;
-    NSString *elementType = dict[[NSNumber numberWithInt:type]];
+    NSString *elementType = dict[[NSNumber numberWithInt:(int)type]];
     if (elementType) {
         return elementType;
     }
